Input automatically zeroes
When the InputMode is secure, the input's clear function will always
overwrite with zeroes first. Additionally, Input now zeroes on drop when
the mode was secure.

Finally, a bug fix in Console::drop that prevented shutting down,
because the strong count was actually 2 due to the frontend thread
having a clone of the state.